requirements-dev.txt
# Development dependencies for running tests and linters.
-r requirements.txt
pymarkdownlnt
pytest
pytest-cov
PyYAML
ruff
yamllint
yq
